requirements.txt
asgiref==3.5.2
captcha==0.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-reversion==5.0.2
django-stubs==1.13.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
djangorestframework-stubs==1.8.0
docopt==0.6.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.991
mypy-extensions==0.4.3
Naked==0.1.31
num2words==0.5.12
packaging==21.3
Pillow==9.2.0
pip-autoremove==0.10.0
psycopg2-binary==2.9.5
pyannotate==1.2.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
transliterate==1.10.2
types-cryptography==3.3.23
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
